+++ x/foundation/msgs.go
@@ -29,6 +29,21 @@ func (m MsgUpdateParams) GetSigners() []sdk.AccAddress {
 	return []sdk.AccAddress{signer}
 }
 
+// Type implements the LegacyMsg.Type method.
+func (m MsgUpdateParams) Type() string {
+	return sdk.MsgTypeURL(&m)
+}
+
+// Route implements the LegacyMsg.Route method.
+func (m MsgUpdateParams) Route() string {
+	return sdk.MsgTypeURL(&m)
+}
+
+// GetSignBytes implements the LegacyMsg.GetSignBytes method.
+func (m MsgUpdateParams) GetSignBytes() []byte {
+	return sdk.MustSortJSON(ModuleCdc.MustMarshalJSON(&m))
+}
+
 var _ sdk.Msg = (*MsgFundTreasury)(nil)
 
 // ValidateBasic implements Msg.
@@ -48,6 +63,21 @@ func (m MsgFundTreasury) ValidateBasic() error {
 func (m MsgFundTreasury) GetSigners() []sdk.AccAddress {
 	signer := sdk.MustAccAddressFromBech32(m.From)
 	return []sdk.AccAddress{signer}
+}
+
+// Type implements the LegacyMsg.Type method.
+func (m MsgFundTreasury) Type() string {
+	return sdk.MsgTypeURL(&m)
+}
+
+// Route implements the LegacyMsg.Route method.
+func (m MsgFundTreasury) Route() string {
+	return sdk.MsgTypeURL(&m)
+}
+
+// GetSignBytes implements the LegacyMsg.GetSignBytes method.
+func (m MsgFundTreasury) GetSignBytes() []byte {
+	return sdk.MustSortJSON(ModuleCdc.MustMarshalJSON(&m))
 }
 
 var _ sdk.Msg = (*MsgWithdrawFromTreasury)(nil)
@@ -75,6 +105,21 @@ func (m MsgWithdrawFromTreasury) GetSigners() []sdk.AccAddress {
 	return []sdk.AccAddress{signer}
 }
 
+// Type implements the LegacyMsg.Type method.
+func (m MsgWithdrawFromTreasury) Type() string {
+	return sdk.MsgTypeURL(&m)
+}
+
+// Route implements the LegacyMsg.Route method.
+func (m MsgWithdrawFromTreasury) Route() string {
+	return sdk.MsgTypeURL(&m)
+}
+
+// GetSignBytes implements the LegacyMsg.GetSignBytes method.
+func (m MsgWithdrawFromTreasury) GetSignBytes() []byte {
+	return sdk.MustSortJSON(ModuleCdc.MustMarshalJSON(&m))
+}
+
 var _ sdk.Msg = (*MsgUpdateMembers)(nil)
 
 // ValidateBasic implements Msg.
@@ -100,6 +145,21 @@ func (m MsgUpdateMembers) GetSigners() []sdk.AccAddress {
 	return []sdk.AccAddress{signer}
 }
 
+// Type implements the LegacyMsg.Type method.
+func (m MsgUpdateMembers) Type() string {
+	return sdk.MsgTypeURL(&m)
+}
+
+// Route implements the LegacyMsg.Route method.
+func (m MsgUpdateMembers) Route() string {
+	return sdk.MsgTypeURL(&m)
+}
+
+// GetSignBytes implements the LegacyMsg.GetSignBytes method.
+func (m MsgUpdateMembers) GetSignBytes() []byte {
+	return sdk.MustSortJSON(ModuleCdc.MustMarshalJSON(&m))
+}
+
 var _ sdk.Msg = (*MsgUpdateDecisionPolicy)(nil)
 
 // ValidateBasic implements Msg.
@@ -122,6 +182,21 @@ func (m MsgUpdateDecisionPolicy) ValidateBasic() error {
 func (m MsgUpdateDecisionPolicy) GetSigners() []sdk.AccAddress {
 	signer := sdk.MustAccAddressFromBech32(m.Authority)
 	return []sdk.AccAddress{signer}
+}
+
+// Type implements the LegacyMsg.Type method.
+func (m MsgUpdateDecisionPolicy) Type() string {
+	return sdk.MsgTypeURL(&m)
+}
+
+// Route implements the LegacyMsg.Route method.
+func (m MsgUpdateDecisionPolicy) Route() string {
+	return sdk.MsgTypeURL(&m)
+}
+
+// GetSignBytes implements the LegacyMsg.GetSignBytes method.
+func (m MsgUpdateDecisionPolicy) GetSignBytes() []byte {
+	return sdk.MustSortJSON(ModuleCdc.MustMarshalJSON(&m))
 }
 
 func (m MsgUpdateDecisionPolicy) GetDecisionPolicy() DecisionPolicy {
@@ -206,6 +281,21 @@ func (m MsgSubmitProposal) GetSigners() []sdk.AccAddress {
 	return signers
 }
 
+// Type implements the LegacyMsg.Type method.
+func (m MsgSubmitProposal) Type() string {
+	return sdk.MsgTypeURL(&m)
+}
+
+// Route implements the LegacyMsg.Route method.
+func (m MsgSubmitProposal) Route() string {
+	return sdk.MsgTypeURL(&m)
+}
+
+// GetSignBytes implements the LegacyMsg.GetSignBytes method.
+func (m MsgSubmitProposal) GetSignBytes() []byte {
+	return sdk.MustSortJSON(ModuleCdc.MustMarshalJSON(&m))
+}
+
 var _ sdk.Msg = (*MsgWithdrawProposal)(nil)
 
 // ValidateBasic implements Msg.
@@ -225,6 +315,21 @@ func (m MsgWithdrawProposal) ValidateBasic() error {
 func (m MsgWithdrawProposal) GetSigners() []sdk.AccAddress {
 	signer := sdk.MustAccAddressFromBech32(m.Address)
 	return []sdk.AccAddress{signer}
+}
+
+// Type implements the LegacyMsg.Type method.
+func (m MsgWithdrawProposal) Type() string {
+	return sdk.MsgTypeURL(&m)
+}
+
+// Route implements the LegacyMsg.Route method.
+func (m MsgWithdrawProposal) Route() string {
+	return sdk.MsgTypeURL(&m)
+}
+
+// GetSignBytes implements the LegacyMsg.GetSignBytes method.
+func (m MsgWithdrawProposal) GetSignBytes() []byte {
+	return sdk.MustSortJSON(ModuleCdc.MustMarshalJSON(&m))
 }
 
 var _ sdk.Msg = (*MsgVote)(nil)
@@ -256,6 +361,21 @@ func (m MsgVote) GetSigners() []sdk.AccAddress {
 	return []sdk.AccAddress{signer}
 }
 
+// Type implements the LegacyMsg.Type method.
+func (m MsgVote) Type() string {
+	return sdk.MsgTypeURL(&m)
+}
+
+// Route implements the LegacyMsg.Route method.
+func (m MsgVote) Route() string {
+	return sdk.MsgTypeURL(&m)
+}
+
+// GetSignBytes implements the LegacyMsg.GetSignBytes method.
+func (m MsgVote) GetSignBytes() []byte {
+	return sdk.MustSortJSON(ModuleCdc.MustMarshalJSON(&m))
+}
+
 var _ sdk.Msg = (*MsgExec)(nil)
 
 // ValidateBasic implements Msg.
@@ -277,6 +397,21 @@ func (m MsgExec) GetSigners() []sdk.AccAddress {
 	return []sdk.AccAddress{signer}
 }
 
+// Type implements the LegacyMsg.Type method.
+func (m MsgExec) Type() string {
+	return sdk.MsgTypeURL(&m)
+}
+
+// Route implements the LegacyMsg.Route method.
+func (m MsgExec) Route() string {
+	return sdk.MsgTypeURL(&m)
+}
+
+// GetSignBytes implements the LegacyMsg.GetSignBytes method.
+func (m MsgExec) GetSignBytes() []byte {
+	return sdk.MustSortJSON(ModuleCdc.MustMarshalJSON(&m))
+}
+
 var _ sdk.Msg = (*MsgLeaveFoundation)(nil)
 
 // ValidateBasic implements Msg.
@@ -292,6 +427,21 @@ func (m MsgLeaveFoundation) ValidateBasic() error {
 func (m MsgLeaveFoundation) GetSigners() []sdk.AccAddress {
 	signer := sdk.MustAccAddressFromBech32(m.Address)
 	return []sdk.AccAddress{signer}
+}
+
+// Type implements the LegacyMsg.Type method.
+func (m MsgLeaveFoundation) Type() string {
+	return sdk.MsgTypeURL(&m)
+}
+
+// Route implements the LegacyMsg.Route method.
+func (m MsgLeaveFoundation) Route() string {
+	return sdk.MsgTypeURL(&m)
+}
+
+// GetSignBytes implements the LegacyMsg.GetSignBytes method.
+func (m MsgLeaveFoundation) GetSignBytes() []byte {
+	return sdk.MustSortJSON(ModuleCdc.MustMarshalJSON(&m))
 }
 
 var _ sdk.Msg = (*MsgGrant)(nil)
@@ -350,6 +500,21 @@ func (m MsgGrant) GetSigners() []sdk.AccAddress {
 	return []sdk.AccAddress{signer}
 }
 
+// Type implements the LegacyMsg.Type method.
+func (m MsgGrant) Type() string {
+	return sdk.MsgTypeURL(&m)
+}
+
+// Route implements the LegacyMsg.Route method.
+func (m MsgGrant) Route() string {
+	return sdk.MsgTypeURL(&m)
+}
+
+// GetSignBytes implements the LegacyMsg.GetSignBytes method.
+func (m MsgGrant) GetSignBytes() []byte {
+	return sdk.MustSortJSON(ModuleCdc.MustMarshalJSON(&m))
+}
+
 var _ sdk.Msg = (*MsgRevoke)(nil)
 
 // ValidateBasic implements Msg.
@@ -375,6 +540,21 @@ func (m MsgRevoke) GetSigners() []sdk.AccAddress {
 	return []sdk.AccAddress{signer}
 }
 
+// Type implements the LegacyMsg.Type method.
+func (m MsgRevoke) Type() string {
+	return sdk.MsgTypeURL(&m)
+}
+
+// Route implements the LegacyMsg.Route method.
+func (m MsgRevoke) Route() string {
+	return sdk.MsgTypeURL(&m)
+}
+
+// GetSignBytes implements the LegacyMsg.GetSignBytes method.
+func (m MsgRevoke) GetSignBytes() []byte {
+	return sdk.MustSortJSON(ModuleCdc.MustMarshalJSON(&m))
+}
+
 var _ sdk.Msg = (*MsgGovMint)(nil)
 
 // ValidateBasic implements Msg.
@@ -398,4 +578,19 @@ func (m MsgGovMint) ValidateBasic() error {
 func (m MsgGovMint) GetSigners() []sdk.AccAddress {
 	signer := sdk.MustAccAddressFromBech32(m.Authority)
 	return []sdk.AccAddress{signer}
+}
+
+// Type implements the LegacyMsg.Type method.
+func (m MsgGovMint) Type() string {
+	return sdk.MsgTypeURL(&m)
+}
+
+// Route implements the LegacyMsg.Route method.
+func (m MsgGovMint) Route() string {
+	return sdk.MsgTypeURL(&m)
+}
+
+// GetSignBytes implements the LegacyMsg.GetSignBytes method.
+func (m MsgGovMint) GetSignBytes() []byte {
+	return sdk.MustSortJSON(ModuleCdc.MustMarshalJSON(&m))
 }